requirements.txt
streamlit
openai
pytube
youtube-transcript-api
serpapi
fpdf
python-dotenv
